Send (sftp/ftp) current file to remote machine
This module takes the currently open file and sends it to a remote machine.
FTP and SFTP are implemented. The connection defaults are stored in the file
and can be overwritten with command-line parameters.

Outsourcing those into a (globa?) config file would be more appropriate.

In the context of this module and the imported pysftp module:
In core/database.py added connect_args={'check_same_thread':False}
to create_engine() parameters, otherwise exceptions were thrown.

According to https://docs.python.org/2/library/sqlite3.html#multithreading
this should be fine.