Add validation of the configuration file
Add a method to the Config class to validate the provided configuration
file. The validation currently checks whether a given set of (statically
defined) sections is present in the config file.
The reason for this PR is that a lot of users most likely use a config
file that is not in the git path (e.g. ~/.viper/viper.conf). When we add
a feature that requires a new sections then some user might not notice
this and will never update their config file.