Add support for project notes
Support for notes that are not associated with a file. A user can create
a new project note by calling the notes command with the flag --add when
no file is open. To create a project file when a file is open, include
the --project flag. It is also possible to list all notes in the
project using the --list flag when no file is open or including the
--project flag when a file is open.